x11: move X11Surface to submodule to reexport and reorganize internal functions
